+++ src/main/java/net/csibio/aird/sample/SampleCode.java
@@ -10,15 +10,14 @@
 
 package net.csibio.aird.sample;
 
+import java.io.File;
+import java.util.List;
+import java.util.TreeMap;
 import net.csibio.aird.AirdManager;
 import net.csibio.aird.bean.AirdInfo;
 import net.csibio.aird.bean.common.Spectrum;
 import net.csibio.aird.parser.DIAParser;
 import net.csibio.aird.util.AirdScanUtil;
-
-import java.io.File;
-import java.util.List;
-import java.util.TreeMap;
 
 /**
  * Sample Code for Aird SDK
@@ -29,43 +28,43 @@ public class SampleCode {
 //        getAirdInfo();
 //    }
 
-    /**
-     * 读取某一个文件夹下所有的Aird Index文件
-     * <p>
-     * Sample code for reading all the aird files in target directory
-     */
-    public static void scanDIAFiles() {
-        List<File> files = AirdScanUtil.scanIndexFiles("E:\\data\\SGS");
-        files.forEach(file -> {
-            try {
-                AirdManager.getInstance().load(file.getPath());
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        });
-    }
+  /**
+   * 读取某一个文件夹下所有的Aird Index文件
+   * <p>
+   * Sample code for reading all the aird files in target directory
+   */
+  public static void scanDIAFiles() {
+    List<File> files = AirdScanUtil.scanIndexFiles("E:\\data\\SGS");
+    files.forEach(file -> {
+      try {
+        AirdManager.getInstance().load(file.getPath());
+      } catch (Exception e) {
+        e.printStackTrace();
+      }
+    });
+  }
 
-    /**
-     * 读取一个Aird文件的元数据文件,并且根据文件中的Block块索引信息读取Aird文件中的光谱图信息 API Samples
-     */
-    public static void getAirdInfo() throws Exception {
-        //Init with the Aird Index(.json) file path
-        DIAParser parser = new DIAParser(
-                "C:\\Users\\LMS\\Desktop\\C20181208yix_HCC_DIA_T_46A_1st.json");
+  /**
+   * 读取一个Aird文件的元数据文件,并且根据文件中的Block块索引信息读取Aird文件中的光谱图信息 API Samples
+   */
+  public static void getAirdInfo() throws Exception {
+    //Init with the Aird Index(.json) file path
+    DIAParser parser = new DIAParser(
+        "C:\\Users\\LMS\\Desktop\\C20181208yix_HCC_DIA_T_46A_1st.json");
 
-        //Read the Aird Index into memory
-        AirdInfo airdInfo = parser.getAirdInfo();
+    //Read the Aird Index into memory
+    AirdInfo airdInfo = parser.getAirdInfo();
 
-        //Read the first spectrum
-        Spectrum pairs = parser.getSpectrum(1);
+    //Read the first spectrum
+    Spectrum pairs = parser.getSpectrum(1);
 
-        //Read SWATH block one by one for DIA type Aird File
-        airdInfo.getIndexList().forEach(blockIndex -> {
-            TreeMap<Float, Spectrum<double[], float[], double[]>> map = parser.getSpectra(blockIndex);
-            System.out.println(map.size());
-        });
+    //Read SWATH block one by one for DIA type Aird File
+    airdInfo.getIndexList().forEach(blockIndex -> {
+      TreeMap<Double, Spectrum<double[], float[], double[]>> map = parser.getSpectra(blockIndex);
+      System.out.println(map.size());
+    });
 
-        //Read TIC
-        airdInfo.getTicMap();
-    }
+    //Read TIC
+    airdInfo.getTicMap();
+  }
 }
